增加获取自增键的配置

+++ src/main/java/net/wanho/mapper/base/BaseMapper.java
@@ -1,16 +1,16 @@
 package net.wanho.mapper.base;
 
 public interface BaseMapper<T> {
-    int deleteByPrimaryKey(Long tid);
+    Long deleteByPrimaryKey(Long tid);
 
-    int insert(T record);
+    Long insert(T record);
 
-    int insertSelective(T record);
+    Long insertSelective(T record);
 
     T selectByPrimaryKey(Long tid);
 
-    int updateByPrimaryKeySelective(T record);
+    Long updateByPrimaryKeySelective(T record);
 
-    int updateByPrimaryKey(T record);
+    Long updateByPrimaryKey(T record);
 
 }
